mobile/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic[email]==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
aiosqlite==0.20.0
greenlet==3.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
opencv-python-headless==4.10.0.84
scikit-image==0.24.0
numpy==1.26.4
Pillow==10.4.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0

# ── ML / CV model dependencies ─────────────────────────────────────────────
# TF CPU wheel — lighter than full tensorflow, no CUDA needed on server
tensorflow-cpu>=2.12.0,<2.16.0
# EasyOCR for denomination digit recognition
easyocr>=1.7.0
# scikit-learn + joblib for BankNote-Net sklearn classifiers
scikit-learn>=1.3.0
